Add GitHub action for running flake8
Add .github/workflows/lint.yml, which lists the GitHub actions, which we
will use for linting. (We may have more linting actions in the same
file, for example pylint, or more actions is separate files)

Inside the file, define `flake8` job, which does the bare minimum
to run flake8 on our codebase:
- checkout the code
- install python
- install flake8
- run flake8

This takes ~23 seconds, compared to the ~3 minutes it takes Travis CI to
run the flake8 task. The difference is that Travis does much more
prerequisites, like installing additional dependencies, which we
don't need in this case.